v0.5.0-alpha: UI & performance improvements, enhanced support for existing features

+++ app/src/main/java/nl/marc_apps/ovgo/data/DepartureRepository.kt
@@ -17,7 +17,7 @@ class DepartureRepository(
             return emptyList()
         }
         val stationsList = trainStationRepository.getTrainStations()
-        val trainInfoList = trainInfoRepository.getTrainInfo(departuresList.map { it.product.number.toInt() }.toSet())
+        val trainInfoList = trainInfoRepository.getTrainInfo(departuresList.map { it.product.number }.toSet())
 
         val converter = ApiDepartureToDomainDeparture(stationsList, resolveTrainInfo = { journeyId ->
             trainInfoList.find { it.journeyId == journeyId.toInt() }
